added length to QuadraticBezierCurve

+++ src/bezierCurve/LinearBezierCurve.java
@@ -1,18 +1,18 @@
 package bezierCurve;
 
 public class LinearBezierCurve {
-    int[] start;
-    int[] end;
+    double[] start;
+    double[] end;
     
-    public LinearBezierCurve(int[] start, int[] end) {
+    public LinearBezierCurve(double[] start, double[] end) {
         this.start = start;
         this.end = end;
     }
 
-    public int[] calculatePosition(double t) {
-        int[] pos = new int[2];
-        pos[0] = (int)((double)(end[0] - start[0])*t) + start[0];
-        pos[1] = (int)((double)(end[1] - start[1])*t) + start[1];
+    public double[] calculatePosition(double t) {
+        double[] pos = new double[2];
+        pos[0] = ((end[0] - start[0])*t) + start[0];
+        pos[1] = ((end[1] - start[1])*t) + start[1];
         return pos;
     }
 }
